feat(rule): 규칙 인터페이스 생성
