+++ src/main/kotlin/domain/state/running/BlackTurn.kt
@@ -1,6 +1,6 @@
 package domain.state.running
 
-import domain.rule.OmokRule
+import domain.rule.RenjuOmokRule
 import domain.state.end.End
 import domain.state.State
 import domain.stone.Board
@@ -10,14 +10,9 @@ import domain.stone.StoneType
 class BlackTurn(board: Board) : Running(board) {
     override fun next(stone: Stone): State {
         if (!isValidPut(stone)) return BlackTurn(board)
-        if (checkForbidden(board, stone)) return BlackTurn(board)
+        if (omokRule.isForbidden(board.board, stone)) return BlackTurn(board)
         board.putStone(stone)
-        if (OmokRule.isWinCondition(board.board, stone)) return End(StoneType.BLACK)
+        if (omokRule.isWinCondition(board.board, stone)) return End(StoneType.BLACK)
         return WhiteTurn(board)
-    }
-
-    private fun checkForbidden(board: Board, stone: Stone): Boolean {
-        return OmokRule.countOpenThrees(board.board, stone) >= OmokRule.MIN_OPEN_THREES ||
-            OmokRule.countOpenFours(board.board, stone) >= OmokRule.MIN_OPEN_FOURS
     }
 }
