+++ src/main/kotlin/domain/state/running/Running.kt
@@ -1,11 +1,15 @@
 package domain.state.running
 
+import domain.rule.OmokRule
+import domain.rule.RenjuOmokRule
 import domain.state.State
 import domain.stone.Board
 import domain.stone.Stone
 import domain.stone.StoneType
 
 abstract class Running(val board: Board) : State {
+
+    val omokRule: OmokRule = RenjuOmokRule()
 
     abstract override fun next(stone: Stone): State
