+++ src/main/kotlin/domain/state/running/WhiteTurn.kt
@@ -1,6 +1,6 @@
 package domain.state.running
 
-import domain.rule.OmokRule
+import domain.rule.RenjuOmokRule
 import domain.state.end.End
 import domain.state.State
 import domain.stone.Board
@@ -11,7 +11,7 @@ class WhiteTurn(board: Board) : Running(board) {
     override fun next(stone: Stone): State {
         if (!isValidPut(stone)) return WhiteTurn(board)
         board.putStone(stone)
-        if (OmokRule.isWinCondition(board.board, stone)) return End(StoneType.WHITE)
+        if (omokRule.isWinCondition(board.board, stone)) return End(StoneType.WHITE)
         return BlackTurn(board)
     }
 }
