+++ src/test/kotlin/domain/rule/RenjuOmokRuleTest.kt
@@ -7,10 +7,11 @@ import domain.stone.StoneType
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
-class OmokRuleTest {
+class RenjuOmokRuleTest {
 
     @Test
     fun `5개이상의 돌이 가로로 이어져 있으면 오목조건을 충족한다`() {
+        val renjuOmokRule: RenjuOmokRule = RenjuOmokRule()
         val board: Board = Board()
         board.putStone(Stone(StonePosition.from(2, 1)!!, StoneType.BLACK))
         board.putStone(Stone(StonePosition.from(3, 1)!!, StoneType.BLACK))
@@ -20,11 +21,12 @@ class OmokRuleTest {
 
         val stone: Stone = Stone(StonePosition.from(6, 1)!!, StoneType.BLACK)
 
-        assertThat(OmokRule.isWinCondition(board.board, stone)).isTrue
+        assertThat(renjuOmokRule.isWinCondition(board.board, stone)).isTrue
     }
 
     @Test
     fun `5개이상의 돌이 세로로 이어져 있으면 오목조건을 충족한다`() {
+        val renjuOmokRule: RenjuOmokRule = RenjuOmokRule()
         val board: Board = Board()
         board.putStone(Stone(StonePosition.from(1, 1)!!, StoneType.BLACK))
         board.putStone(Stone(StonePosition.from(1, 2)!!, StoneType.BLACK))
@@ -34,11 +36,12 @@ class OmokRuleTest {
 
         val stone: Stone = Stone(StonePosition.from(1, 5)!!, StoneType.BLACK)
 
-        assertThat(OmokRule.isWinCondition(board.board, stone)).isTrue
+        assertThat(renjuOmokRule.isWinCondition(board.board, stone)).isTrue
     }
 
     @Test
     fun `5개이상의 돌이 오른쪽 대각선으로 이어져 있으면 오목조건을 충족한다`() {
+        val renjuOmokRule: RenjuOmokRule = RenjuOmokRule()
         val board: Board = Board()
         board.putStone(Stone(StonePosition.from(1, 1)!!, StoneType.BLACK))
         board.putStone(Stone(StonePosition.from(2, 2)!!, StoneType.BLACK))
@@ -48,11 +51,12 @@ class OmokRuleTest {
 
         val stone: Stone = Stone(StonePosition.from(5, 5)!!, StoneType.BLACK)
 
-        assertThat(OmokRule.isWinCondition(board.board, stone)).isTrue
+        assertThat(renjuOmokRule.isWinCondition(board.board, stone)).isTrue
     }
 
     @Test
     fun `5개이상의 돌이 왼쪽 대각선으로 이어져 있으면 오목조건을 충족한다`() {
+        val renjuOmokRule: RenjuOmokRule = RenjuOmokRule()
         val board: Board = Board()
         board.putStone(Stone(StonePosition.from(1, 5)!!, StoneType.BLACK))
         board.putStone(Stone(StonePosition.from(2, 4)!!, StoneType.BLACK))
@@ -62,6 +66,6 @@ class OmokRuleTest {
 
         val stone: Stone = Stone(StonePosition.from(5, 1)!!, StoneType.BLACK)
 
-        assertThat(OmokRule.isWinCondition(board.board, stone)).isTrue
+        assertThat(renjuOmokRule.isWinCondition(board.board, stone)).isTrue
     }
 }
